pyproject.toml
[tool.poetry]
name = "10-favorit-empl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
aiogram = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
